Refactor course card component to use ng-container for conditional rendering
The update makes use of the ng-container directive to handle the *ngIf structural directive for conditional rendering of the course card component. This helps in grouping of DOM elements without influencing styles or layout (extra 'div' elements won't be created, the page will be more lightweight). The category level has also been wrapped with the ng-container for conditional rendering based on course level.